product controller

+++ routes/api.php
@@ -13,42 +13,45 @@ Route::middleware(['auth:sanctum'])->get('/user', function (Request $request) {
 });
 
 
-Route::name('api.')
-    ->prefix('v1')
+Route::name('api.v1')//va a scrivere nella prima parte del name->route
+    ->prefix('v1')//va a inserirsi nel link del  sito
+//     ->middleware(['auth:sanctum']) 
     ->group(function () {
     ////////////////// ROUTE PRODUCT /////////////////////
         Route::get('/products',            [ProductController::class, 'index'])->name('products.index');// http://127.0.0.1:8000/api/v1/products
         Route::get('/products/{product}',  [ProductController::class, 'show'])->name('products.show');//http://127.0.0.1:8000/api/v1/products/id
 
 
-        Route::get('/products/add',     [ProductController::class, 'add'])->name('products.add');
-        Route::put('/products/{id}/edit',  [ProductController::class, 'update'])->name('products.edit');
-        Route::delete('/products/{id}',    [ProductController::class, 'delete'])->name('products.delete');
+        Route::post('/products/add',     [ProductController::class, 'add'])->name('products.add');//ADMIN
+        Route::put('/products/{id}/edit',  [ProductController::class, 'edit'])->name('products.edit');//ADMIN
+        Route::delete('/products/{id}',    [ProductController::class, 'delete'])->name('products.delete');//ADMIN
 
-
+///////// FINO QUI TUTTO CORRETTO ////////
 //////////////////// ROUTE ORDER /////////////////////
 
         Route::get('/orders',            [OrderController::class, 'index'])->name('orders.index'); //http://127.0.0.1:8000/api/v1/orders
         Route::get('/orders/{order}',     [OrderController::class, 'show'])->name('orders.show');
-        Route::get('/orders/confirm',     [OrderController::class, 'confirm'])->name('orders.confirm');
-        Route::delete('/products/{id}',    [ProductController::class, 'delete'])->name('products.delete');
+
+        Route::get('/orders/confirm',     [OrderController::class, 'confirm'])->name('orders.confirm');//ADMIN
+        Route::delete('/products/{id}',    [ProductController::class, 'delete'])->name('products.delete');//ADMIN
 
 
 
 
 //////////////////// ROUTE CART /////////////////////
 
-        Route::get('/cart',            [CartController::class, 'index'])->name('cart');///carrello utente
+        Route::get('/cart',            [CartController::class, 'index'])->name('cart');//CLIENT
 
-        Route::get('/carts/add',       [CartController::class, 'create'])->name('carts.add');
-        Route::delete('/carts/{id}',   [CartController::class, 'delete'])->name('carts.delete');
-        Route::get('/carts/',          [CartController::class, 'buy'])->name('carts.buy');
+        Route::get('/carts/add',       [CartController::class, 'add'])->name('carts.add'); //CLIENT
+
+        Route::delete('/carts/{id}',   [CartController::class, 'delete'])->name('carts.delete'); //CLIENT
+        Route::get('/carts/',          [CartController::class, 'buy'])->name('carts.buy');//CLIENT
 
 //////////////////// ROUTE REVIEW /////////////////////
 
-        Route::get('/reviews',            [ReviewController::class, 'index'])->name('reviews.index');
-        Route::get('/reviews/add',       [ReviewController::class, 'add'])->name('reviews.add');
-        Route::delete('/reviews/{id}',    [ReviewController::class, 'delete'])->name('reviews.delete');
+        Route::get('/reviews',            [ReviewController::class, 'index'])->name('reviews.index');//TUTTI
+        Route::get('/reviews/add',       [ReviewController::class, 'add'])->name('reviews.add');//CLIENT
+        Route::delete('/reviews/{id}',    [ReviewController::class, 'delete'])->name('reviews.delete');//CLIENT
 
 //////////////////// ROUTE TYPE USER /////////////////////
